+++ commons/ava_services/gecko_service_initializer_core_test.go
@@ -3,17 +3,15 @@ package ava_services
 import (
 	"bytes"
 	"fmt"
+	"net"
 	"testing"
 
 	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_services/cert_providers"
 	"github.com/kurtosis-tech/kurtosis/commons/services"
 	"github.com/stretchr/testify/assert"
-	"net"
 )
 
-
 var testPublicIp = net.ParseIP("172.17.0.2")
-
 
 func TestNoDepsStartCommand(t *testing.T) {
 	initializerCore := NewGeckoServiceInitializerCore(
@@ -30,7 +28,7 @@ func TestNoDepsStartCommand(t *testing.T) {
 		"--public-ip=" + testPublicIp.String(),
 		"--network-id=local",
 		"--http-port=9650",
-		"--http-host=" + testPublicIp.String(),
+		"--http-host=",
 		"--staking-port=9651",
 		"--log-level=info",
 		"--snow-sample-size=1",
@@ -63,7 +61,7 @@ func TestWithDepsStartCommand(t *testing.T) {
 		"--public-ip=" + testPublicIp.String(),
 		"--network-id=local",
 		"--http-port=9650",
-		"--http-host=" + testPublicIp.String(),
+		"--http-host=",
 		"--staking-port=9651",
 		"--log-level=info",
 		"--snow-sample-size=1",