+++ gecko_client/health_api.go
@@ -20,11 +20,11 @@ func (api HealthApi) GetLiveness() (health.GetLivenessReply, error) {
 	var response health.GetLivenessReply
 	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(healthApiEndpoint, "health.getLiveness", make(map[string]interface{}))
 	if err != nil {
-		return response, stacktrace.Propagate(err, "Error getting liveness")
+		return health.GetLivenessReply{}, stacktrace.Propagate(err, "Error getting liveness")
 	}
 
 	if err := json.Unmarshal(responseBodyBytes, &response); err != nil {
-		return response, stacktrace.Propagate(err, "Error unmarshalling JSON response")
+		return health.GetLivenessReply{}, stacktrace.Propagate(err, "Error unmarshalling JSON response")
 	}
 	return response, nil
 }