+++ gecko_client/health_api_test.go
@@ -4,31 +4,55 @@ import (
 	"testing"
 )
 
+type healthTest struct {
+	resultString    string
+	expectedChecks  int
+	expectedHealthy bool
+	nilError        bool
+}
+
 func TestGetLiveness(t *testing.T) {
-	// resultStr := `{
-	//    "jsonrpc":"2.0",
-	//    "result":{
-	//       "checks":{
-	//          "network.validators.heartbeat":{
-	//             "message":{
-	//                "heartbeat":1591041377
-	//             },
-	//             "timestamp":"2020-06-01T15:56:18.554202-04:00",
-	//             "duration":23201,
-	//             "contiguousFailures":0,
-	//             "timeOfFirstFailure":null
-	//          }
-	//       },
-	//       "healthy":true
-	//    },
-	//    "id":1
-	// }`
+	tests := []healthTest{
+		{
+			resultString: `{
+            "jsonrpc":"2.0",
+            "result":{
+               "checks":{
+                  "network.validators.heartbeat":{
+                     "message":{
+                        "heartbeat":1591041377
+                     },
+                     "timestamp":"2020-06-01T15:56:18.554202-04:00",
+                     "duration":23201,
+                     "contiguousFailures":0,
+                     "timeOfFirstFailure":null
+                  }
+               },
+               "healthy":true
+            },
+            "id":1
+         }`,
+			expectedChecks:  1,
+			expectedHealthy: true,
+			nilError:        true,
+		},
+		// TODO fix health API to not error when parsing health check that contains an error
+	}
 
-	// client := clientFromRequester(mockedJsonRpcRequester{resultStr: resultStr})
-	// livenessInfo, err := client.HealthApi().GetLiveness()
-	// assert.Nil(t, err, "Error message should be nil")
-	// assert.True(t, livenessInfo.Healthy)
+	for _, test := range tests {
+		client := clientFromRequester(mockedJsonRpcRequester{resultStr: test.resultString})
+		livenessInfo, err := client.HealthApi().GetLiveness()
 
-	// checks := livenessInfo.Checks
-	// assert.Equal(t, len(checks), 1)
+		if test.nilError && err != nil {
+			t.Fatalf("Expected error to be nil, but found: %w", err)
+		}
+
+		if test.expectedChecks != len(livenessInfo.Checks) {
+			t.Fatalf("Expected to find: %d checks, but instead found %d checks", test.expectedChecks, len(livenessInfo.Checks))
+		}
+
+		if test.expectedHealthy && !livenessInfo.Healthy {
+			t.Fatal("Unexpectedly returned unhealthy")
+		}
+	}
 }